requirements.txt
qiskit-terra== 0.16.4
qiskit-aer==0.7.6
qiskit-ignis==0.5.2
qiskit-ibmq-provider==0.12.1
qiskit-aqua==0.8.2
qiskit==0.24.0
shapely==1.7.1
gurobipy==9.1.2
cplex ==20.1.0.0
xlrd==2.0.1
xlwt==1.3.0
openpyxl==3.0.7
